fix(core): validate project names at instance creation
Duplicate project names now fail at createRstest or CLI startup regardless of --project selection. The check runs in resolveRunnerInputs on all declared projects, and the dead creation-time filterProjects call in resolveProjects is removed because --project is applied per operation.
