+++ e2e/programmatic/index.test.ts
@@ -12,22 +12,6 @@ const parsePayload = (stdout: string) =>
   parseMarkerPayload<Record<string, any>>(stdout, '__RSTEST_API_RESULT__');
 
 describe('programmatic createRstest', () => {
-  it('validates duplicate names only among selected projects', async ({
-    onTestFinished,
-  }) => {
-    const { cli } = await runRstestCli({
-      command: 'node',
-      args: ['run-duplicate-projects.mjs'],
-      onTestFinished,
-      options: { nodeOptions: { cwd: fixturesDir } },
-    });
-    await cli.exec;
-    const result = parsePayload(cli.stdout);
-    expect(result.status).toBe('pass');
-    expect(result.passed).toBe(1);
-    expect(result.error).toContain('Project name "beta" is already used');
-  });
-
   it.for(['forks', 'vmThreads'] as const)(
     'uses inline config and rejects build failures with %s',
     async (pool, { onTestFinished }) => {
