--- e2e/projects/fixtures-duplicate-names/alpha/index.test.js
@@ -1,3 +1,0 @@
-import { test } from '@rstest/core';
-
-test('works', () => {});
